Convert TPC row generators to hooks js_only
